* todo_beta/refactor/mint_allocation_dao:
  refactor: consolidate float params to float64

# Conflicts:
#	x/tokenomics/keeper/msg_server_update_param.go
#	x/tokenomics/types/message_update_param.go

+++ pkg/client/interface.go
@@ -330,7 +330,7 @@ type BlockQueryClient interface {
 // protobuf message. Since the generated go types don't include interface types, this
 // is necessary to prevent dependency cycles.
 type ProofParams interface {
-	GetProofRequestProbability() float32
+	GetProofRequestProbability() float64
 	GetProofRequirementThreshold() *cosmostypes.Coin
 	GetProofMissingPenalty() *cosmostypes.Coin
 	GetProofSubmissionFee() *cosmostypes.Coin
